change a few typos on html.  change payback year to double,  display resutls in doubleboxes

+++ com.blackout.solarpanelcalculator/src/com/blackout/solarpanelcalculator/client/CalculationService.java
@@ -7,5 +7,5 @@ import com.google.gwt.user.client.rpc.RemoteServiceRelativePath;
 public interface CalculationService extends RemoteService {
 	double doDailySolarGeneration(double systemSize, double roofEfficiency, double inverterEfficiency, double wiringEfficiency, double whatYear, double agingEfficiencyLoss, double solarIrradiance);
 	double doDailySavings(double dailyGeneration, double exportPercent, double replacePercent, double feedInTarrif, double powerCost);
-	double doPayBackYear(double systemCost, int lifeSpan, double dailySavings);
+	double doPayBackYear(double systemCost, double lifeSpan, double dailySavings);
 }
